If storage.conf graphdriver=="", devmapper shouldn't be tried
This patch removes devicemapper from the list of drivers tried when driver=""

This change is necessary since tests in tools that use containers storage are
being run when driver is not being specified. In this situation the test fails
because devmapper returns and error other then ErrNotSupported.   With this
change a useful message will go to the logs if the user actually specifies
"devicemapper" in config.

Signed-off-by: Daniel J Walsh <dwalsh@redhat.com>