Add some tests for mapping using images
Add tests to exercise both handling mapping at the per-container level,
and at the in-image-layer level.  While having the image manage top
layers is preferable because it's faster for cases where multiple
containers which use the same ID mappings are being created from a
single image, for read-only image stores, that's not an option, so we
need to still be capable of doing things in the container's layer.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>