sysregistriesv2: forbid URI schemes
Forbid the usage of URI schemes (e.g., http, https, etc.) in the URLs of
registries and mirrors.  The reasoning behind is that the `Insecure`
option is sufficient to control skipping of the certs verification and
to allow non-TLS connections.

The change of forbidding URI schemes entails some changes in the code:

- Move from net/url.URL to simple strings as the parsing does not work
  properly when URI schemes aren't present.

- The rather complex embedding of different types to allow
  (un)marshalling can now be fully avoided.

- Update the tests accordingly.

Refer to the following issue for a background discussion:
    https://github.com/projectatomic/libpod/issues/721

Signed-off-by: Valentin Rothberg <vrothberg@suse.com>