Add manifest.Schema1.ExtractedV1Compatibility
This allows us to simplify consumers of the data, and notably will allow
future LayerInfos to work without having to worry about parsing errors.

manifest.Schema1FromManifest is parsing all of that data anyway in
fixManifestLayers, so this is a net win for those manifests; OTOH
manifest.Schema1FromComponents may do the work unnecessarily.

Signed-off-by: Miloslav Trmač <mitr@redhat.com>