Add EmptyLayer to manifest.LayerInfo.
This makes the "empty layer" flag available for storageImage.

Also continues to add tests to the manifest subpackage.

A possible alternative was to add this right to types.BlobInfo,
but that exists primarily for copying images around, where
the consumers don't care about the value, and producers
like PutBlob would have to do extra work to preserve the (unused)
value; whether the value is or isn't available in a particular
instance would have to be only documented in text.

This allows us to use the type system to verify availability
of the flag: if you want it, call manifest.LayerInfos explicitly.
(Of course if it turns out to be more widely needed, we can undo this
and move the flag directly to BlobInfo.)

Signed-off-by: Miloslav Trmač <mitr@redhat.com>