docker tarfile: support auto-decompression of source
This matches how "docker load" deals with compressed images, as well as
being a general quality-of-life improvement over the previous error
messages we'd give. This also necessarily removes the previous
special-cased gzip handling, and adds support for auto-decompression for
streams as well.

Signed-off-by: Aleksa Sarai <asarai@suse.de>

(Updated primarily to handle closing the decompressed streams.)
Signed-off-by: Miloslav Trmač <mitr@redhat.com>