Add blobinfocache.NoCache
This exists primarily for the benefit of Image.ConfigBlob
and similar readers, which don't really benefit from the DiffID matching,
blob mounting, and the like, or callers like ImageDestination.PutBlob,
which actively wants to avoid using any extra cached data (because the
cache lookup should have already happened and now it would just be
redundant).

Does not add any users yet, so should not change behavior.

Signed-off-by: Miloslav Trmač <mitr@redhat.com>