Don't update manifests when writing to digested references
Don't make changes to manifests or the blobs they reference when we're
writing to a location that's specified using a digested reference, since
changing the manifest will break the reference.

Enforce this by checking if our destination reference includes a digest,
and if it does, refusing to copy an image whose manifest doesn't already
match it.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>