WIP: Support multiple manifests in c/storage
When writing an image, always record the manifest in a "manifest-$digest"
item, in addition to the existing "manifest" item.

To make this consistent, fail in PutManifest if the destination reference
includes a digest and the manifest does not match.

When reading an image, return the manifest corresponding to the user-specified
digest (if such a manifest has been recorded), falling back to the existing
use of "manifest" (i.e. the latest one written).

Signed-off-by: Miloslav Trmač <mitr@redhat.com>