Try harder in storageImageDestination.TryReusingBlob
Because a storage.Layer only records a single CompressedDigest
and CompressedSize, we may need to depend on the BlobInfoCache.

In that case, TryReusingBlob currently returns correct data
about the uncompressed blob, causing the manifest ultimately sent
to PutManifest to be updated, and potentially breaking digest references,
or disabling the reuse if manifest updates are forbidden.

The extra information necessary to avoid the manifest update is
just the compressed layer size, and with non-schema1 manifests that
size is a part of the manifest and provided by the caller; so, we can
just return the caller's BlobInfo unmodified.

To make this work, Commit also needs to be updated to call
LayersByUncompressedDigest on a diffID recorded in blobDiffIDs
because the kinds of layers we are discussing here can't be found by
LayersByCompressedDigest.

Signed-off-by: Miloslav Trmač <mitr@redhat.com>