Fallback to v2s2 if OCI manifest upload fails on ECR
Uploading a spec-compliant OCI manifest which contains no top-level
media type fails on AWS ECR with an Unsupported error code and
an 'Invalid JSON syntax' error message. Catch this case as an
invalid manifest error so that we automatically fallback to other
accepted media types in copyOneImage().

Addresses libpod issue #1719

Signed-off-by: Mark McLoughlin <markmc@redhat.com>