Use separate short-term progress pools for layers, and config
Use a short-term progress pool for the parallel layer copy in copyLayers,
and a single-use progress pool in copyConfig.

Then drop the long-term copier.progressPool, which was being incorrectly
deinitialized on the _first_ execution of copyUpdatedConfigAndManifest
(making uses of the pool in future executions, if schema conversion is
necessary, invalid).

For now, only call progressPool.Wait() when the code succesfully
finishes copies, because progressPool.Wait() would currently block
until all bars in the pool are "completed" (by SetTotal, by reading
the expected amount of data, or by cancellation).

Signed-off-by: Miloslav Trmač <mitr@redhat.com>