README: Drop '**' from "Contributing" header
These are from 35aa6fa9 (Make it easier for CONTRIBUTORS to find
information, 2018-07-05, #474).  I dunno if they were intentional or
not, but I thought they might be footnote markers so I scrolled down
to the bottom of the README looking for a footnote.  Remove them to
save other people from that same confusion ;).

Signed-off-by: W. Trevor King <wking@tremily.us>