Remove unfinished VFS layer if its creation fails.
Current implementation of VFS driver performs the copy of the parent
layer during the creation of new (child) layer, but does not clean
the destination if the copy process fails.

From the consuming application perspective, this make storage leaks
possible, e.g. an attempt to create new container with podman,
if failed in the middle of the copy due to disk overflow, leaves
unfinished container layer stray.

To avoid obliging the application (or the end user) to take care of
such possible leaks, cleanup is added.

Signed-off-by: Danila Kiver <danila.kiver@mail.ru>