docker client: error if registry is blocked
Return an error whenever we attempt to create a docker client when the
underlying registry is configured to be blocked in the registries.conf.
Until now, users of the c/image library implemented this behaviour but
it has not been enforced.

Now, it's time to pull the trigger and enforce the configuration.

Signed-off-by: Valentin Rothberg <rothberg@redhat.com>