We should not be touching internal work dir
Vivek Goyal says that we should not be touching the internal WorkDir/work directory.
This is internal to the kernel and should not be modified.  I believe this was done
to make sure that the "/" of the container is writable by the root user of the user
namespace that is running the container.  Changing the ownership of the diffDir
should give us the same behavior and is safe to do.

Signed-off-by: Daniel J Walsh <dwalsh@redhat.com>