Add tests for parsing OpenShift kubeconfig files
github.com/imdario/mergo has changed behavior incompatibly;
verify that the parsing works as expected.

The kubeconfig fixture contains invalid data for certificates and private
keys to make the current tests easier to read - but that also prevents
testing restClientFor and its callers which actually need the data.
Right now I prefer readable tests for the parts we care about, eventually
the fixture might be updated with real data - but, ultimately, that is
already tested in Skopeo integration tests, so testing it here as well
is not that important (considering the atomic: transport is deprecated
already).

Signed-off-by: Miloslav Trmač <mitr@redhat.com>