Revert "update dependencies"
This reverts commit fcd86c4c681071def3606ddcb606adda592f3e1a, which
was mostly _reverting_ dependencies.

(The reason for that commit was that mergo has incompatibly changed
behavior, causing atomic: cluster accesses to use localhost:8080
instead of the cluster API endpoint defined in the configuration.
The previous two commits have updated mergo and adjusted the callers,
so we no longer need to hold mergo back.)

Signed-off-by: Miloslav Trmač <mitr@redhat.com>