Check if the layer is actually mounted.
There are cases where the storage database gets out of whack with
whether or not the storage is actually mounted.  We need to check
before returning the mount point.

1 A user could go in and umount the storage.
2 If the storage was mounted in a different mount namespace and then
the mount namespace goes away the counter will never get decremented
even though the mount point was removed.
3. If storage runtime is on non tmpfs storage a system reboot could be
done that will not clear the mount count.

This patch will fix the problem with the layer not being mounted, but
we still have a problem in that we can't figure out when to umount the
image. Not sure that is a solveable problem.

Signed-off-by: Daniel J Walsh <dwalsh@redhat.com>