Initial Code of Conduct
This is the initial Code of Conduct document for use across
all of the projects under the github.com/containers repository.

The projects will each link to this document from each of their
repos.

This document was pulled from libpod which had pulled it from
Kubernetes who had pulled it from the CNCF.

Signed-off-by: TomSweeneyRedHat <tsweeney@redhat.com>