determineListConversion: check forced type
If determineListConversion() is given a forced type, change how we
decide which type to use by treating that value as the lone supported
type.  This will reject unrecognized manifest types which we would have
quietly accepted here, only to fail to convert them later.

Add some debug messages so that we know what determineListConversion()
decided, even if its decision was forced.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>